Deduplicate generated types.
This introduces a concept of a `TypeId` that can be used to deduplicate
types that are output.  This is loosely based on the built in rust
`TypeId`, but with some additions to support types for variants (which
can't have a `TypeId` in rust)

I've put some effort into making sure the order of output is the same,
so this shouldn't change in this commit.  Though may want to change it
in the future.

Fixes #3